pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "topoflow"
version = "1.0.0"
description = "Physics-informed deep learning for multi-pollutant air quality forecasting"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ammar Kheder"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
keywords = [
    "air quality",
    "deep learning",
    "transformer",
    "atmospheric science",
    "weather prediction",
    "physics-informed",
]
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "timm>=0.9.0",
    "numpy>=1.20.0",
    "xarray>=2023.0.0",
    "zarr>=2.14.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/ammarkheder/topoflow"
Repository = "https://github.com/ammarkheder/topoflow"
Documentation = "https://github.com/ammarkheder/topoflow#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["topoflow*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]
